refactor: Ghost 모델과 플레이어 이동 로직 분리
-fix: 클라이언트 only 테스트 시 로그인 체크 과정에서 서버가 오류나면 사이트가 통째로 죽는 버그 수정
